chore: clean up / styling adjustments in new toast components
